@@ -5,11 +5,22 @@ import { SignUpComponent } from './sign-up/sign-up.component';
 import { HomeComponent } from './home/home.component';
 // import { AppComponent } from './app.component';
 import { AuthenticateComponent } from './authenticate/authenticate.component';
+import { AuthGuard } from './auth-guard.guard';
+import { LogoutComponent } from './logout/logout.component';
+import { ScheduleComponent } from './schedule/schedule.component';
+import { ErrComponent } from './err/err.component';
 
 export const routes: Routes = [
   { path: '', component: HomeComponent },
-  { path: 'Login', component: LoginComponent },
+  { path: 'login', component: LoginComponent },
+  { path: 'logout', component: LogoutComponent },
+  { path: 'schedule', component: ScheduleComponent },
   { path: 'signUp', component: SignUpComponent },
-  { path: 'authenticate/:activePage', component: AuthenticateComponent },
-  { path: '**', redirectTo: '' }
+  { path: 'err', component: ErrComponent },
+  {
+    path: 'authenticate',
+    component: AuthenticateComponent,
+    canActivate: [AuthGuard]
+  },
+  { path: '**', redirectTo: 'err' }
 ];
