@@ -1,5 +1,26 @@
-import { CanActivateFn } from '@angular/router';
+import { Injectable } from '@angular/core';
+import {
+  Router,
+  CanActivate,
+  ActivatedRouteSnapshot,
+  RouterStateSnapshot
+} from '@angular/router';
 
-export const authGuardGuard: CanActivateFn = (route, state) => {
-  return true;
-};
+import { AccountService } from './auth-service';
+
+@Injectable({ providedIn: 'root' })
+export class AuthGuard implements CanActivate {
+  constructor(private router: Router, private accountService: AccountService) {}
+
+  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
+    const user = this.accountService.isRegistring;
+    if (user) {
+      // authorised so return true
+      return true;
+    }
+
+    // not logged in so redirect to login page with the return url
+    this.router.navigate(['/err']);
+    return false;
+  }
+}
