@@ -1,5 +1,7 @@
-import { Component } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { RouterLink, RouterLinkActive, RouterOutlet } from '@angular/router';
+
+import { AccountService } from '../auth-service';
 
 @Component({
   selector: 'app-header-component',
@@ -8,13 +10,31 @@ import { RouterLink, RouterLinkActive, RouterOutlet } from '@angular/router';
   templateUrl: './header.component.html',
   styleUrl: './header.component.css'
 })
-export class HeaderComponent {
+export class HeaderComponent implements OnInit {
+  loggedIn: boolean = false;
+  user?: string | null;
+  private accountService = inject(AccountService);
 
- active: boolean = false;
+  ngOnInit(): void {
+    this.accountService.isLoggedIn$.subscribe((data: any) => {
+      if (data) {
+        this.loggedIn = true;
+        // this.user = data.email;
+      } else {
+        this.loggedIn = false;
+      }
+    });
+    this.accountService.loggedEmail$.subscribe((data: any) => {
+      if (data) {
+        console.log(data);
+        this.user = data;
+      } else {
+        this.user = null;
+      }
+    });
+  }
 
-//  onActive() {
-//   this.active = true;
-//   return true;
-//  }
-
+  onLogout() {
+    this.accountService.logOut();
+  }
 }
