@@ -1,4 +1,4 @@
-import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import {
   AfterViewInit,
   Component,
@@ -10,7 +10,6 @@ import {
   ViewChild
 } from '@angular/core';
 import {
-  AbstractControl,
   FormControl,
   FormGroup,
   ReactiveFormsModule,
@@ -18,7 +17,9 @@ import {
   Validators
 } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
-import { catchError, EMPTY, fromEvent, of, Subscription } from 'rxjs';
+import { catchError, fromEvent, of, Subscription } from 'rxjs';
+import { AccountService } from '../auth-service';
+// import { HTTPService } from '../http-service';
 
 type objA = {
   message: string;
@@ -28,17 +29,20 @@ type objA = {
 @Component({
   selector: 'app-sign-up',
   standalone: true,
-  imports: [HttpClientModule, ReactiveFormsModule],
+  imports: [ReactiveFormsModule],
   templateUrl: './sign-up.component.html',
   styleUrl: './sign-up.component.css'
 })
 export class SignUpComponent
   implements OnInit, AfterViewInit, OnDestroy, OnChanges
 {
+  private accountService = inject(AccountService);
+  private route = inject(ActivatedRoute);
   private http = inject(HttpClient);
-  private clickedEmail: Subscription = new Subscription();
-  private clickedPassword: Subscription = new Subscription();
-  private clickedConfirmPassword: Subscription = new Subscription();
+  private router = inject(Router);
+
+  private clickedElement = new Subscription();
+
   currentPage?: string;
 
   signUpForm!: FormGroup;
@@ -60,9 +64,7 @@ export class SignUpComponent
   @ViewChild('pwd', { static: false }) el1?: ElementRef;
   @ViewChild('cpwd', { static: false }) el2?: ElementRef;
 
-  private route = inject(ActivatedRoute);
-
-  constructor(private router: Router) {
+  constructor() {
     this.signUpForm = new FormGroup({
       'email': new FormControl(null, [Validators.email, Validators.required]),
       'password': new FormControl(null, [
@@ -76,11 +78,6 @@ export class SignUpComponent
 
   private date?: string;
   ngOnInit(): void {
-    // this.route.paramMap.subscribe((params) => {
-    //   this.currentPage = params.get('activePage')!;
-    //   console.log(this.currentPage);
-    // });
-
     this.route.queryParams.subscribe((params) => {
       this.date = params['activePage'];
       console.log(this.date); // Print the parameter to the console.
@@ -90,69 +87,65 @@ export class SignUpComponent
   ngOnChanges(): void {}
 
   ngAfterViewInit(): void {
-    this.clickedEmail = fromEvent(
-      this.el0!.nativeElement,
-      'focusout'
-    ).subscribe(() => {
-      this.onCheckEmail(this.el0?.nativeElement.value)
-        .pipe(catchError((err) => of([])))
-        .subscribe((data: any) => {
-          this.emailTouched = true;
-          this.canUseEmail = data.canUseEmail;
-          if (!this.canUseEmail) {
-            this.checkEmailMsg = data.message;
-          } else if (!this.validateEmail) {
-            this.checkEmailMsg = 'Please enter a valid e-mail';
-          } else if (this.canUseEmail && this.validateEmail) {
-            this.checkEmailMsg = 'E-mail is available';
-          } else {
-            this.checkEmailMsg = '';
-          }
-        });
-      this.validateEmail =
-        this.signUpForm!.get('email')?.valid &&
-        this.signUpForm!.get('email')?.touched;
-    });
+    this.clickedElement.add(
+      fromEvent(this.el0!.nativeElement, 'focusout').subscribe(() => {
+        this.onCheckEmail(this.el0?.nativeElement.value);
+      })
+    );
 
-    this.clickedPassword = fromEvent(
-      this.el1!.nativeElement,
-      'focusout'
-    ).subscribe(() => {
-      this.passwordTouched = true;
-      this.validatePassword =
-        this.signUpForm!.get('password')?.valid &&
-        this.signUpForm!.get('email')?.touched;
-    });
+    this.clickedElement.add(
+      fromEvent(this.el1!.nativeElement, 'focusout').subscribe(() => {
+        this.passwordTouched = true;
+        this.validatePassword =
+          this.signUpForm!.get('password')?.valid &&
+          this.signUpForm!.get('email')?.touched;
+      })
+    );
 
-    this.clickedConfirmPassword = fromEvent(
-      this.el2!.nativeElement,
-      'focusout'
-    ).subscribe(() => {
-      const password = this.signUpForm!.get('confirmpassword')?.value;
-      const cpassword = this.signUpForm!.get('password')?.value;
-      this.cpasswordTouched = true;
-      if (
-        password === cpassword &&
-        this.signUpForm!.get('confirmpassword')?.touched
-      ) {
-        this.validateConfirmPassword = true;
-      } else {
-        this.validateConfirmPassword = false;
-      }
-    });
+    this.clickedElement.add(
+      fromEvent(this.el2!.nativeElement, 'focusout').subscribe(() => {
+        const password = this.signUpForm!.get('confirmpassword')?.value;
+        const cpassword = this.signUpForm!.get('password')?.value;
+        this.cpasswordTouched = true;
+        if (
+          password === cpassword &&
+          this.signUpForm!.get('confirmpassword')?.touched
+        ) {
+          this.validateConfirmPassword = true;
+        } else {
+          this.validateConfirmPassword = false;
+        }
+      })
+    );
   }
 
   ngOnDestroy(): void {
-    this.clickedEmail.unsubscribe();
-    this.clickedPassword.unsubscribe();
-    this.clickedConfirmPassword.unsubscribe();
+    this.clickedElement.unsubscribe();
   }
 
   onCheckEmail(email: string) {
     this.email = email;
-    return this.http.post('http://localhost:8080/signUp/checkEmail', {
-      email: this.email
-    });
+    this.http
+      .post('http://localhost:8080/signUp/checkEmail', {
+        email: this.email
+      })
+      .pipe(catchError((err) => of([])))
+      .subscribe((data: any) => {
+        this.emailTouched = true;
+        this.canUseEmail = data.canUseEmail;
+        if (!this.canUseEmail) {
+          this.checkEmailMsg = data.message;
+        } else if (!this.validateEmail) {
+          this.checkEmailMsg = 'Please enter a valid e-mail';
+        } else if (this.canUseEmail && this.validateEmail) {
+          this.checkEmailMsg = 'E-mail is available';
+        } else {
+          this.checkEmailMsg = '';
+        }
+      });
+    this.validateEmail =
+      this.signUpForm!.get('email')?.valid &&
+      this.signUpForm!.get('email')?.touched;
   }
 
   onSubmit() {
@@ -171,43 +164,16 @@ export class SignUpComponent
         password: this.password
       })
       .subscribe((data: any) => {
-        if (data.message !== 'User created') {
-          console.log('there is an error');
+        if (data.message !== `User ${this.email} created`) {
+          console.log(data.verify);
         }
-        const url = Date.now();
-        this.router.navigate(['/authenticate', data.token]);
+        this.accountService.Registring = true;
+        this.router.navigate(['/authenticate'], {
+          queryParams: { email: this.email }
+        });
       });
     this.el0!.nativeElement.value = '';
     this.el1!.nativeElement.value = '';
     this.el2!.nativeElement.value = '';
   }
-
-  // private passwordMatch(
-  //   password: string,
-  //   confirmPassword: string
-  // ): ValidatorFn {
-  //   return (formGroup: AbstractControl): { [key: string]: any } | null => {
-  //     const passwordControl = formGroup.get(password);
-  //     const confirmPasswordControl = formGroup.get(confirmPassword);
-
-  //     if (!passwordControl || !confirmPasswordControl) {
-  //       return null;
-  //     }
-
-  //     if (
-  //       confirmPasswordControl.errors &&
-  //       !confirmPasswordControl.errors['passwordMismatch']
-  //     ) {
-  //       return null;
-  //     }
-
-  //     if (passwordControl.value !== confirmPasswordControl.value) {
-  //       confirmPasswordControl.setErrors({ passwordMismatch: true });
-  //       return { passwordMismatch: true };
-  //     } else {
-  //       confirmPasswordControl.setErrors(null);
-  //       return null;
-  //     }
-  //   };
-  // }
 }
